# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GhostProtocol dApp API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GhostState {
  /** Total drift score */
  zeta: number;
  /** Phase accumulation */
  phi: number;
  /** Quarantine probability (0-1) */
  omega: number;
  /** Coupling divergence (0-1) */
  coupling: number;
  locked: boolean;
  compromised: boolean;
  escaped: boolean;
  forked: boolean;
  anomalyCount: number;
  evidenceCounter: number;
  totalRewardsPaid: number;
  isQuarantined: boolean;
  scoreYou: number;
  scoreHer: number;
  paused: boolean;
}

export interface GhostMetrics {
  marketCap: number;
  price: number;
  priceChange24h: number;
  volume24h: number;
  holders: number;
  totalSupply: number;
  /** Percentage of people who have ghosted someone */
  ghostingIndex: number;
  /** Global ghosting to rejection ratio */
  rejectionRate: number;
  /** Community-aggregated heartbreak level 0-100 */
  heartbreakScore: number;
  /** Simulated counter of unread texts */
  readReceiptsIgnored: number;
}

export type GhostEventType =
  (typeof GhostEventType)[keyof typeof GhostEventType];

export const GhostEventType = {
  DriftUpdated: "DriftUpdated",
  PhiUpdated: "PhiUpdated",
  Compromised: "Compromised",
  Locked: "Locked",
  Escaped: "Escaped",
  Forked: "Forked",
  EvidenceAdded: "EvidenceAdded",
  Anomaly: "Anomaly",
  GhostGained: "GhostGained",
} as const;

export interface GhostEvent {
  id: string;
  type: GhostEventType;
  timestamp: string;
  message: string;
  value?: number;
}

export interface HolderEntry {
  rank: number;
  address: string;
  alias: string;
  balance: number;
  percentage: number;
  lastActivity: string;
}

export interface EvidenceSubmission {
  /** Evidence hash (bytes32 as hex string) */
  hash: string;
  /** Evidence weight */
  weight: number;
  /** Human readable description */
  description?: string;
}

export interface EvidenceResult {
  success: boolean;
  evidenceCounter: number;
  message: string;
  newZeta: number;
}