# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GhostProtocol dApp API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ghost/submit-evidence:
    post:
      operationId: submitEvidence
      tags: [ghost]
      summary: Submit evidence to the ghost protocol
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvidenceSubmission"
      responses:
        "200":
          description: Evidence submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvidenceResult"
components:
  schemas:
...
    EvidenceSubmission:
      type: object
      properties:
        hash:
          type: string
          description: Evidence hash (bytes32 as hex string)
        weight:
          type: number
          description: Evidence weight
        description:
          type: string
          description: Human readable description
      required:
        - hash
        - weight
    EvidenceResult:
      type: object
      properties:
        success:
          type: boolean
        evidenceCounter:
          type: number
        message:
          type: string
        newZeta:
          type: number
      required:
        - success
        - evidenceCounter
        - message
        - newZeta