# lib/api-zod/src/generated/types/evidenceResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GhostProtocol dApp API
 * OpenAPI spec version: 0.1.0
 */

export interface EvidenceResult {
  success: boolean;
  evidenceCounter: number;
  message: string;
  newZeta: number;
}