# lib/api-zod/src/generated/types/evidenceSubmission.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * GhostProtocol dApp API
 * OpenAPI spec version: 0.1.0
 */

export interface EvidenceSubmission {
  /** Evidence hash (bytes32 as hex string) */
  hash: string;
  /** Evidence weight */
  weight: number;
  /** Human readable description */
  description?: string;
}